klatt.cpp: fix a sound spike in setzeroabc when f == 0.
If setzeroabc is passed f == 0, rp->a will be set to 0 and when
the coefficients get inverted (a'=1/a, b'=b/a, c'=c/a) then a',
b' and c' get set to INF. This results in an audible sound spike
when these resonators get used.

Found when investigating an issue in klatt with example1.par when
using the cascade synthesis mode.

Bug: https://github.com/rhdunn/klatt/issues/1.