Make eSpeak-specific setting changes apply immediately.
This reverts commit 433a0a04556de9959b98bbc0afb8c351886e22d2.

The issue here appears to be that the process the eSpeak commands are
processed from to generate the speech differs from the one used to
modify the settings. This appears to be enough to cause it to not
get the updated settings until the process is restarted.

Thanks to Peter Vagner for locating the cause.