SeekBarPreference: don't make the value text selectable
The ADT lint warnings flagged the value text as "consider making
the text selectable", however if you do TalkBack will cause the
first character to be read every time the seekbar changes poisiton
(due to the text being updated via the setText method).

This is annoying and confusing, so textIsSelectable is set to false
to supress the lint warning.