SeekBarPreference: update the persisted value when the SeekBar changes value
The idea here is that whenever the user changes one of the parameter
values via the SeekBar control, it is reflected immediately by eSpeak.
This is similar to how e.g. the brightness preference UI works.
However, instead of updating every time the value changes while the
slider is being moved, it is updated once the touch interaction is
finished.

Also, if the user Cancels or exits the dialog with the back button,
the old setting value is restored.