Fix a SEGFAULT in GetTranslatedPhonemeString.
When running any of the following:

    espeak -x "[[]]"
    espeak --ipa "[[]]"

eSpeak would SEGFAULT. This is because phon_out_buf is NULL as no
phonemes have been written and as this is the first call to
GetTranslatedPhonemeString, the phoneme buffer has not been
allocated yet. However, GetTranslatedPhonemeString attempts to add
a NULL character to the end of the phoneme buffer, and thus crashes.