wxwidgets 3.0: move FONT_* construction to SpectFrame
When using wxWidgets 3.0, creating a wxFont requires the gtk+
library to be initialized. This is causing espeakedit to segfault.

SpectFrame::Draw is the only place where the FONT_* variables are
used, so move them into the SpectFrame class and create them in
the constructor.

Also, FONT_NORMAL is not used so it has been removed.