Always pass plist to WritePhMnemonic [clang scan-build]
The clang scan-build static analysis tool warns of a null pointer
dereference in WritePhMnemonic. This is only for phonSWITCH codes,
which is guarded against in the calls that pass NULL to plist, so
the null dereference should not trigger.

Always passing plist supresses this warning, and prevents any
future errors.