issue 172: fix sluggish speech for stopped speech.
This restores the behaviour removed in commit
ca831d236a4eed5648254329554affac143a1740.

When using eSpeak for audio and eSpeak in a server, using the async
speech requests, that code stopped the audio as soon as possible
from a cancel request. This allows servers like brltty to be more
responsive.