Ensure phoneme type is set before start/endtype.
The eSpeak code uses `end_type` for the `voicingswitch` property
when used with consonants. It also allows `endtype` to be specified
in consonants if it is the same as `starttype`.

This code adds a check in the `starttype` and `endtype` handlers
to ensure that the phoneme type has been set to preserve the
eSpeak behaviour for these properties.