Don't save an intonation if it was not found.
Clang static analysis reports an 'Assigned value is garbage or
undefined' error. This happens if the default_tune data has not
been copied in the kTUNE case.

Extending this further, it does not make sense to save the tune
data if the tune was not found, or if the tune already exists.