Copy name to tr->dictionary_name if not equal
This is a similar change to b60d2452c34ac6ebf01a3c09c17193b8c8e2a3fd.
In this case, it is when tr->dictionary_name is passed as the name
parameter in LoadDictionary.

This happens in the SetTranslator2 function when loading the
dictionary for the second language translator object.