en: don't velarize n across word boundaries
It is possible -- especially at higher speeds -- for the n at the
end of a word to be velarised if the next word starts with a velar
plosive. I prefer the velarised sound between word boundaries, but
others do not. As such, limit the velarisation to within the word
only.

[1]	https://en.wikipedia.org/wiki/English_phonology