Add word "summe" in rules for correct compound word pronunciation
I hope that will be my last change. Most of them, like Zitat, gedulds, etc have been added for correct compound word pronunciation. The A: phoneme in komma and beta sounds more natural for me; I also did Komma this way because it matches the way espeak reads it when written as an actual word.