Added 2 rules and made 1 change to an existing one
The first rule I added was to indicate that "ur" at the end of a word is stressed so that it now can pronounce words like Gravur, Tastatur, Klausur, Zensur, etc. correctly. The second rule I added was a word which it pronounced incorrectly before. The change to the existing rules were so that it pronounces kanal and signal  everywhere with the stress on the a.