sl: remove the Cyrillic to Latin replacements.
These were added in espeak 1.47.11a, but Slovenian uses Latin
like Polish and Czech.

The earliest form of Slovene (the Freising Manuscripts, between
972-1039) are written in Latin.

Some Cyrillic and Cyrillic-like letters were used in addition to
Latin in the Dajnko and Metelko alphabets between 1824 and 1838,
but these alphabets did not catch on.

The modern form is based on the Serbo-Croatian Gaj's Latin alphabet.
This is the alphabet represented in the espeak `replace_cyrillic_latin`
table. Slovenian does not use the Cyrillic variant of this alphabet.