Update VoiceVariantPreference.java
Hi, I'm a fan of this project, I don't know if pasting the complete code of the file in an edit box respects all the indentation and structure of the same file but I just wanted to contribute to the project with the addition of new voice variants to the pre-existing on Android ... Could someone please check this out? Anyway, I would expect new voices in the next EspeakNG compilations from FDroid or Google play.