Workround for #945; check for 'ru sum strings'
The change made to 'tests/translate.test' to detect bug #824 also
demonstrates what has been described to me as a compiler dependency in
the output; on my system every occurence of "a " in the phoneme output
is actually output as "a# ".

This is reported in issue #945 (I could find no prior report) however
the issue does not seem to be fixed anytime soon; it has been around
since the test was added in 30214437fc0bb0d067bb60cd550e192edcd2a626
on Dec. 20, 2020.

The test terminates 'make check' in the case in question and therefore
obscures any following errors and makes it very difficult to debug
unrelated changes to any part of espeak-ng or the build system.

This change adds support for tests/common:test_phon for a MESSAGE
argument "ru sum strings" which was added to the test_phon call by
17e6bd0672421467554dcca95f04c2a63b70c510 on the 16th inst. (although
that change doesn't seem to do anything).

If "ru sum strings" is passed every occurence of "a# " in the output
received from espeak-ng is changed to "a ".  Since "a# " does not occur
in the correct output this makes no difference to the original check.

Obviously I've only tested this on my system, I know the issue is known
but since there did not seem to be a bug report until I entered #945 I
don't know what output other systems generate (other than the correct
output).  The change is intended to allow "make check" to skip this
known error and therefore detect other issues.

Signed-off-by: John Bowler <jbowler@acm.org>