Isolate 'Translate' buffers
A number of buffers are used temporarily during translate.  Removing
information from the header files ensures that this is true and allows
these buffers to be more easily managed.  This is a move towards that;
like the previous changes in this series it is just moving stuff around
in a way that should be completely safe.

Testsed gcc-11.1.0 -Wall -Wextra -Werror and make check.

Signed-off-by: John Bowler <jbowler@acm.org>