Make PHONEME_LIST2 and related arrays dynamic
This fixes the #845/#945 memory overwrite issue and is a partial step to
a full MT capable version (i.e. a version with no global memory).  The
arrays are extended incrementally via realloc, the algorithm is crude
(to the point of nonexistence), the arrays start off a fixed length and
get stepped up each time by a fixed increment, however because the
arrays are retained across calls using the same Translator there is some
hope that they will be adaptive within a single language stream -
LoadDictionary destroys the Translator so the arrays are reset by a
dictionary change.

Signed-off-by: John Bowler <jbowler@acm.org>