Full fixes for #945, #847 and maybe others
This fixes the memory overwrite documented in #945 and #847.  It also
fixes a large variety of changes to the phoneme stream that happen as a
result of the same overwrite and were dependent of the layout of global
memory, therefore highly compiler dependent.

The change consists primarily of the changes to translate.c in the
previous commit; this commit includes corrections to those changes and
further large-text test results.

The bug arose because of incorrect checking of the index into the global
'ph_list2' and because of phoneme (ph_list2) truncations or elisions
that occured when the checks correctly detected array overflow.  Rather
than trying to fix the checks up; the whole code relied on guesstimated
fixed sizes for static arrays, this fix dynamically allocates all the
arrays in question.

Take note that the fix does not dynamically allocate the final phoneme
list, which is still fixed size so therefore might get truncated (but it
did before) and, while it does dynamically allocate the input 'source'
buffer it does not dynamically re-allocate it.  This means that the
current behavior of inserting a spurious pause if the clause exceeds the
limits of the input buffer is preserved.  This is necessary; it does
seem to prevent truncation of the final phoneme list.

This is quite deliberately a partial fix: a full fix would dynamically
re-allocate the input buffer within ReadClause and dynamically allocate
the output phoneme list (MakePhonemeList).  Nevertheless these changes
can be provided incrementally; the advantage of this approach is that
substantial code changes only occur in translate.c and are sufficient on
their own to fix the bugs.

The changes have been tested with gcc-8.4.0 and gcc-11.1.0 using -O2
compilation.  Many builds have also been done with -O0 while debugging.
The test sequence is:

git clean -dfx
bash -xe autogen.sh
CC=<compiler> ./configure --prefix=<installdir> [OPTS]
make all
make check
make install

Then further tests were run with the installed espeak-ng.  Initially no
extra options were provided.  In the final tests --with-extdict-* were
added to add all three extended dictionaries.

Because the original bug was produced by an arbitrary long test further
tests were run based on real-world texts in three languages:

KingJamesBible en-gb-x-rp https://www.gutenberg.org/ebooks/10
紅樓夢 cmn-latn-pinyin https://www.gutenberg.org/ebooks/24264
ВОМНАиМИРТэ ru https://archive.org/details/voinaimir00tols

The first entry is a tag used in the timing figures below, the second is
the test language and the third is the origin of the test text; bare
UTF-8 in all cases.  Note that the War and Peace is tricky; the 'text'
download actually downloads an HTML page.  I used this URL, quoted in
other pages referring to the archived text, to obtain an undecorated
form:

ihttps://ia600904.us.archive.org/1/items/voinaimir00tols/voinaimir00tols_djvu.txt

To validate these tests I compared the output phonemes using 'espeak
-xq' between the offical HEAD at the time,
40b78f5183f66958b333d8a54f0ab25a9576be5e and this commit.  I validated
output with both GCC 8.4.0 and 11.1.0.

I also validated overall performance using these texts.  Espeak-ng is
single threaded and the tests show substantially all time is in user
mode, at least on my system.  The following table shows user-mode time
in seconds for before/after with each compiler expressed as a percentage
of the unchanged
8.4.0 time:

Test            11.1.0 +fix:   8.4.0  11.1.0
ВОМНАиМИРТэ    100.418       100.269 100.523
紅樓夢         100.363       100.308 100.484
KingJamesBible 100.291       100.227 100.608

It can be seen that there is no significant change in performance for
large texts.

Signed-off-by: John Bowler <jbowler@acm.org>