Add Quenya ⟨q⟩ as equivalent of ⟨qu⟩
According to the Tolkien Gateway article on Qenya [1], Tolkien used ⟨q⟩
to represent [kʷ] before the writing of The Lord of the Rings (during
which he decided to use ⟨qu⟩ instead). Let’s support it, in case someone
wants to use eSpeak NG to pronounce names from the earlier volumes of
the History of Middle-earth.

[1]: http://tolkiengateway.net/wiki/Qenya